refactor: slim root layout - extract overlays and notification routing
- GlobalOverlays component absorbs the ThemedAlert/Toast/update/changelog/
  signal-bubble JSX
- useNotificationRouting hook owns notification response handling, FCM
  foreground bridging, and the pending-route queue
- deduplicate the deferred-startup block and the clear-history-reminder
  toast logic (each appeared twice)

_layout.tsx: 611 -> 420 lines, 19 responsibilities down to ~9
